Split different modal activators onto modules
Problem: The code in modals.js is too big to maintain.
Solution: Split it onto modules.
